requirements.txt
fastapi
uvicorn
bcrypt==4.0.1
python-jose
passlib[bcrypt]
python-multipart
pydantic
python-dotenv
motor
pydantic[email]
pytesseract
pillow
opencv-python
numpy
httpx
apscheduler